LayoutTests/imported/w3c/web-platform-tests/html/canvas/element/manual/drawing-images-to-the-canvas/drawimage_svg_image_1-expected.txt


FAIL Load a 100x100 image to a SVG image and draw it to a 100x100 canvas. assert_equals: Red channel of the pixel at (0, 0) expected 253 but got 255
